almost there :)
git-svn-id: https://darktable.svn.sourceforge.net/svnroot/darktable@96 2165daf7-ed08-4abd-b3da-b2de91a31a1b

+++ src/develop/develop.h
@@ -53,8 +53,8 @@ typedef struct dt_develop_t
   GList *iop;
 
   // histogram for display.
-  uint32_t *histogram, *histogram_pre;
-  uint32_t histogram_max, histogram_pre_max;
+  float *histogram, *histogram_pre;
+  float histogram_max, histogram_pre_max;
   uint8_t gamma[0x100];
 }
 dt_develop_t;
