+++ src/gui/histogram.h
@@ -17,6 +17,6 @@ gboolean dt_gui_histogram_motion_notify(GtkWidget *widget, GdkEventMotion *event
 gboolean dt_gui_histogram_button_press(GtkWidget *widget, GdkEventButton *event, gpointer user_data);
 gboolean dt_gui_histogram_button_release(GtkWidget *widget, GdkEventButton *event, gpointer user_data);
 gboolean dt_gui_histogram_leave_notify(GtkWidget *widget, GdkEventCrossing *event, gpointer user_data);
-void dt_gui_histogram_draw_8(cairo_t *cr, uint32_t *hist, int32_t channel);
+void dt_gui_histogram_draw_8(cairo_t *cr, float *hist, int32_t channel);
 
 #endif
